// src/api/vcci-news/models/advertisement.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { AdvertisementFile } from './advertisementFile';
import type { AdvertisementType } from './advertisementType';
import type { AdvertisementStatus } from './advertisementStatus';

export interface Advertisement {
  id: string;
  /** @maxLength 255 */
  name: string;
  file_id: string;
  /** @nullable */
  file?: AdvertisementFile;
  /**
   * @maxLength 255
   * @nullable
   */
  alt?: string | null;
  /** @maxLength 500 */
  link: string;
  type: AdvertisementType;
  status: AdvertisementStatus;
  sort_order: number;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  created_by?: string | null;
  /** @nullable */
  updated_at?: string | null;
  /** @nullable */
  updated_by?: string | null;
}