// src/api/vcci-news/models/advertisementCreate.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { AdvertisementCreateType } from './advertisementCreateType';
import type { AdvertisementCreateStatus } from './advertisementCreateStatus';

export interface AdvertisementCreate {
  /** @maxLength 255 */
  name: string;
  file_id: string;
  /**
   * @maxLength 255
   * @nullable
   */
  alt?: string | null;
  /** @maxLength 500 */
  link: string;
  type?: AdvertisementCreateType;
  status?: AdvertisementCreateStatus;
  sort_order?: number;
}