// src/api/vcci-news/models/advertisementUpdate.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { AdvertisementUpdateType } from './advertisementUpdateType';
import type { AdvertisementUpdateStatus } from './advertisementUpdateStatus';

export interface AdvertisementUpdate {
  /** @maxLength 255 */
  name?: string;
  file_id?: string;
  /**
   * @maxLength 255
   * @nullable
   */
  alt?: string | null;
  /** @maxLength 500 */
  link?: string;
  type?: AdvertisementUpdateType;
  status?: AdvertisementUpdateStatus;
  sort_order?: number;
}